Fix HAPI string handling, session bugs and platform defaults
Correctness:

- Util::Hapi::get_string() allocated a buffer without checking whether
  GetStringBufLength succeeded and never terminated it, so a failed length
  query built a std::string out of uninitialised memory. get_parm_string() in
  get_parms.cpp had the same shape with an uninitialised length, which sized
  the allocation from a garbage value. Both go through one checked reader now,
  as do the status, connection error and composed cook result reads.
- Starting a TCP session parsed the port with std::stoi. The extension is built
  without exceptions, so a non-numeric connection string terminated the process
  rather than throwing. Parsed with strtol and range checked now.
- restart_session() ignored its session_type argument and restarted with the
  default named pipe. It now restarts the requested type on the pipe or port
  the previous session used.
- start_session() cast an arbitrary int from GDScript straight to SessionType.
  Unsupported values are rejected with a message, and the three supported types
  are bound as HEGoAPI.SESSION_IN_PROCESS / SESSION_NAMED_PIPE /
  SESSION_TCP_SOCKET so callers stop passing bare numbers.
- The Houdini install fallback was a hardcoded Windows path in three places, so
  Linux and macOS fell back to a path that cannot exist. There is one
  platform-aware HEGoPlatform::get_default_houdini_path() now, and hego.gd warns
  when the configured path is missing - a project.godot authored on Windows
  carries its path to every other machine.
- dlopen and dlclose failures discarded dlerror(), so a missing or unloadable
  libHAPIL surfaced only as a generic later error. Both are logged.

Consistency:

- 14 std::cout/cerr calls in the session manager, the platform layer and the
  HOUDINI_CHECK_ERROR_GET macro now use HEGo::Util::Log like the rest of the
  codebase; a GDExtension's stdout is not where users look.
- Raw new[]/delete[] arrays in hego_api.cpp and get_parms.cpp replaced with
  std::vector, and the hand-rolled HAPI string reads replaced with the shared
  helper.
- Unreal-style leftovers (myCookOptions, bSuccess, SessionResult) renamed to
  match the surrounding snake_case.
- Fixed the wrong include guard comment in hego_session_manager.h.

Also: cook polling no longer spins a core (2 ms sleep between HAPI status
polls), dead unreachable code in load_lib_hapil() removed, and the three
compiler warnings the sources produced under -Wall -Wextra -Wshadow are gone.

Documented what was unobvious: what SessionType values mean and that InProcess
is served by a shared memory server, that the Existing* types are unimplemented,
and where the per-platform Houdini defaults have to stay in sync.

Verified with a clean Linux build.

+++ src/util/geo/fetch_surfaces.cpp
@@ -237,7 +237,7 @@ godot::Dictionary fetch_surfaces(HEGoSessionManager *session_mgr, HAPI_NodeId no
 	godot::Dictionary split_prim_dictionary = HEGo::Util::Geo::build_nested_dictionary(split_attribs, split_attribs_dict, int_ids, read_attribs_dict, 0);
 
 	godot::Array vt_pt_indices;
-	for (int i = 0; i < vertex_point_indices.size(); i++)
+	for (size_t i = 0; i < vertex_point_indices.size(); i++)
 	{
 		vt_pt_indices.append(vertex_point_indices[i]);
 	}
